Issue #99: Refactored `ProcessesList` module as es6, added some extra type definitions.
