--- bw_matchbox/assets/js/ProcessesList/ProcessesListConstants.js
@@ -1,37 +1,0 @@
-modules.define(
-  'ProcessesListConstants',
-  [
-    // Required modules...
-  ],
-  function provide_ProcessesListConstants(
-    provide,
-    // Resolved modules...
-  ) {
-    // Define module...
-
-    /* @desc Shared constants
-     */
-
-    const ProcessesListConstants = {
-      __id: 'ProcessesListConstants',
-
-      // DEBUG: useDebug -- specify debug mode. Don't use it for production!
-      useDebug: false,
-
-      /** Api base */
-      processesApiUrl: '/processes',
-
-      /** The number of records to retrieve at once and to display */
-      pageSize: 25,
-      /** Default order value */
-      defaultOrderBy: 'random',
-      /** Default filter value */
-      defaultFilterBy: 'none',
-      /** Default userDb value */
-      defaultUserDb: 'source', // `Source` is the same as the server provided data
-    };
-
-    // Provide module...
-    provide(ProcessesListConstants);
-  },
-);
