+++ bw_matchbox/assets/modules/ProcessesList/@types/TFilterBy.d.ts
@@ -1,0 +1,2 @@
+type TFilterProcessesBy = 'matched' | 'unmatched' | 'waitlist';
+type TFilterProcessesByString = 'none' | TFilterProcessesBy;
