+++ bw_matchbox/assets/modules/ProcessesList/@types/TOrderBy.d.ts
@@ -1,0 +1,1 @@
+type TOrderByString = 'random' | 'name' | 'location' | 'product';
