+++ bw_matchbox/assets/modules/ProcessesList/@types/TUserDb.d.ts
@@ -1,0 +1,2 @@
+type TUserDb = 'source' | 'target' | 'proxy';
+type TUserDbString = 'none' | TUserDb;
