+++ setup.cfg
@@ -63,6 +63,7 @@ install_requires =
     levenshtein
     tomli
     werkzeug
+    # faker # NOTE: May be used in debug mode (to generate fake comments for tests)
 
 [options.packages.find]
 where = .
